User page component added, has image
